ci: add package smoke coverage
Add an explicit package smoke script that validates built package ESM and CJS entrypoints through package exports, plus subpath type resolution from a temporary consumer project.

Run the smoke script after the package build in the aggregate check script and PR test workflow so build, declarations, and package export regressions are caught before publish.
