+++ .changeset/pr-build-smoke.md
@@ -1,0 +1,5 @@
+---
+"superformdata": patch
+---
+
+Add explicit PR CI coverage for package builds and built package import/type smoke tests.
