+++ .github/workflows/test.yml
@@ -23,3 +23,24 @@ jobs:
 
       - name: Run test suite
         run: bun test
+
+  package-smoke:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+
+      - name: Setup Bun
+        uses: oven-sh/setup-bun@v2
+        with:
+          bun-version: 1.3.7
+
+      - name: Install dependencies
+        run: bun install --frozen-lockfile
+
+      - name: Build package
+        run: bun run build
+
+      - name: Run package smoke tests
+        run: bun run smoke:package
